header для авторизованого користувача
